+++ lib/src/variant/xiangqi.dart
@@ -1,5 +1,7 @@
 part of 'variant.dart';
 
+/// Xiangqi / Cờ Tướng / Chinese Chess defintions.
+/// Still a work in progress, some things are incomplete.
 class Xiangqi {
   static const String defaultFen =
       'rheakaehr/9/1c5c1/p1p1p1p1p/9/9/P1P1P1P1P/1C5C1/9/RHEAKAEHR w - - 0 1';