+++ lib/src/variant/xiangqi.dart
@@ -4,7 +4,7 @@ part of 'variant.dart';
 /// Still a work in progress, some things are incomplete.
 class Xiangqi {
   static const String defaultFen =
-      'rheakaehr/9/1c5c1/p1p1p1p1p/9/9/P1P1P1P1P/1C5C1/9/RHEAKAEHR w - - 0 1';
+      'rnbakabnr/9/1c5c1/p1p1p1p1p/9/9/P1P1P1P1P/1C5C1/9/RNBAKABNR w - - 0 1';
   static PieceType general() => PieceType.fromBetza(
         'W',
         royal: true,
@@ -36,8 +36,8 @@ class Xiangqi {
         pieceTypes: {
           'K': general(),
           'A': advisor(),
-          'E': elephant(),
-          'H': horse(),
+          'B': elephant(),
+          'N': horse(),
           'R': chariot(),
           'C': cannon(),
           'P': soldier(),